Multithreading and Recursion
Implemented multi-threading using the threading facility of the lego-VM.
with a few tweaks to make the implementation as unobtrusive as possible
to a Basic programmer.
Recursion is implemented by putting all basic SUBs together with the
main program in a common VM object and handle calls and returns with a
custom return stack.